Move rebuild_cache and clear_cache calls to SystemController
