+++ open-api/swagger.yaml
@@ -2222,7 +2222,7 @@ paths:
                 numberOfUpdatedSets:
                   type: integer
 
-  /v2/tree_nodes/clear_cache:
+  /v2/admin/system/clear_cache:
     get:
       description: |
         Clears tree node and counts caches.
@@ -2237,13 +2237,11 @@ paths:
           description: |
             The cache has been cleared.
 
-  /v2/tree_nodes/rebuild_cache:
+  /v2/admin/system/rebuild_cache:
     get:
       description: |
         Clears tree node and counts caches and rebuilds the tree node cache per user.
         Only for administrators.
-        Deprecated. Use `/v2/admin/system/after_data_loading_update` instead.
-      deprecated: true
       tags:
         - v2
         - admin
@@ -2259,7 +2257,8 @@ paths:
     get:
       description: |
         This endpoint should be called after loading, deleting or updating data in the database.
-        It does different cache levels recalculations. Clears tree nodes, counts and patient set ids caches. Refreshes the bitsets materialized view. Updates data for subscribed user queries.
+        It does different cache levels recalculations. Clears tree nodes, counts and patient set ids caches. Rebuilds the tree node cache per user.
+        Refreshes the bitsets materialized view. Updates data for subscribed user queries.
         Only for administrators.
       tags:
         - v2
